changed data manager interface to take longs instead of ints for removing contacts

+++ src/com/android/smap/di/DataManager.java
@@ -15,6 +15,6 @@ public interface DataManager {
 
 	void putContacts(List<Contact> contacts);
 
-	void removeContactFromSurvey(int contact, int survey);
+	void removeContactFromSurvey(long contact, long survey);
 
 }
